order-service/requirements.txt
africastalking==1.2.9
aiofiles==24.1.0
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
async-property==0.2.2
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.2
cryptography==45.0.2
deprecation==2.1.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
drf-spectacular==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
josepy==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jwcrypto==1.5.6
kombu==5.5.3
mozilla-django-oidc==4.0.1
packaging==25.0
pika==1.3.2
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-keycloak==5.5.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
schema==0.7.7
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
redis>=4.0.0
celery[redis]>=5.0.0
